VMware SDk integration - refactoring. Signed-off-by: Mihail-Cristian Tusa <mihail-cristian.tusa@hpe.com>
