--- score-vmware/src/main/java/io/cloudslang/content/constants/Constants.java
@@ -1,59 +1,0 @@
-package io.cloudslang.content.constants;
-
-/**
- * Created by Mihai Tusa.
- * 11/13/2015.
- */
-public class Constants {
-
-    // VMware specific constants
-    public static final String SERVICE_INSTANCE = "ServiceInstance";
-    public static final String HOST_SYSTEM = "HostSystem";
-    public static final String DATA_CENTER = "Datacenter";
-    public static final String RESOURCE_POOL = "resourcePool";
-    public static final String ENVIRONMENT_BROWSER = "environmentBrowser";
-    public static final String CONTAINER_VIEW = "ContainerView";
-    public static final String PERSISTENT = "persistent";
-    public static final String GENERATED = "generated";
-    public static final String PARENT = "parent";
-    public static final String STATE = "state";
-    public static final String READY = "ready";
-    public static final String ERROR = "error";
-    public static final String NAME = "name";
-    public static final String VIEW = "view";
-    public static final String INFO_STATE = "info.state";
-    public static final String INFO_ERROR = "info.error";
-    public static final String TEST_CD_ISO = "testcd.iso";
-    public static final String DEFAULT_FLOPPY_DEVICE_NAME = "/dev/fd0";
-    public static final String FILTER_VALUES = "filtervals";
-    public static final String DEFAULT_VOLUME_NAME = "[Local]";
-    public static final String KEY_VALUE_NULL_STRING = "val: null";
-    public static final String SSL = "SSL";
-    public static final String URI_PATH = "/sdk";
-    public static final String VM_FOLDER = "vmFolder";
-    public static final String VIRTUAL_MACHINE = "VirtualMachine";
-    public static final String TRAVERSE_ENTITIES = "traverseEntities";
-
-    public static final int DEFAULT_HTTPS_PORT = 443;
-    public static final int DEFAULT_CPU_COUNT = 1;
-    public static final int DEFAULT_FLOPPY_DEVICE_KEY = 3;
-    public static final int DEFAULT_DISK_KEY = 0;
-    public static final int DEFAULT_DISK_CONTROLLER_KEY = 1;
-    public static final int DEFAULT_DISK_UNIT_NUMBER = 0;
-    public static final int DEFAULT_CDROM_KEY = 20;
-    public static final int DEFAULT_CDROM_UNIT_NUMBER = 0;
-    public static final int DEFAULT_CONTROLLER_BUS_NUMBER = 0;
-    public static final int DEFAULT_NIC_KEY = 4;
-    public static final int THOUSAND_MULTIPLIER = 1000;
-    public static final int MAX_TRIED_WAIT_FOR_UPDATE_COUNTER = 100;
-    public static final int THIRTY = 30;
-    public static final int SIXTY = 60;
-
-    public static final long DEFAULT_VM_DISK_SIZE_MB = 1024;
-    public static final long DEFAULT_VM_MEMORY_SIZE_MB = 1024;
-
-    // Generic constants
-    public static final String EMPTY = "";
-    public static final String LEFT_SQUARE_BRACKET = "[";
-    public static final String RIGHT_SQUARE_BRACKET = "]";
-}
