--- score-vmware/src/main/java/io/cloudslang/content/constants/Outputs.java
@@ -1,20 +1,0 @@
-package io.cloudslang.content.constants;
-
-/**
- * Created by Mihai Tusa.
- * 10/20/2015.
- */
-public class Outputs {
-    // Outputs
-    public static final String RETURN_CODE = "returnCode";
-    public static final String EXCEPTION = "exception";
-    public static final String RETURN_RESULT = "returnResult";
-
-    // Responses
-    public static final String SUCCESS = "success";
-    public static final String FAILURE = "failure";
-
-    // Return codes
-    public static final String RETURN_CODE_SUCCESS = "0";
-    public static final String RETURN_CODE_FAILURE = "-1";
-}
