+++ score-vmware/src/main/java/io/cloudslang/content/vmware/constants/Inputs.java
@@ -1,0 +1,26 @@
+package io.cloudslang.content.vmware.constants;
+
+/**
+ * Created by Mihai Tusa.
+ * 10/20/2015.
+ */
+public class Inputs {
+    // Http client specific inputs
+    public static final String HOST = "host";
+    public static final String PORT = "port";
+    public static final String PROTOCOL = "protocol";
+    public static final String USERNAME = "username";
+    public static final String PASSWORD = "password";
+    public static final String TRUST_EVERYONE = "trustEveryone";
+
+    // VM specific inputs
+    public static final String DATA_CENTER_NAME = "dataCenterName";
+    public static final String HOSTNAME = "hostname";
+    public static final String VIRTUAL_MACHINE_NAME = "virtualMachineName";
+    public static final String VIRTUAL_MACHINE_DESCRIPTION = "description";
+    public static final String DATA_STORE = "dataStore";
+    public static final String CPU_COUNT = "numCPUs";
+    public static final String VM_DISK_SIZE = "vmDiskSize";
+    public static final String VM_MEMORY_SIZE = "vmMemorySize";
+    public static final String GUEST_OS_ID = "guestOsId";
+}
